backend/requirements.txt
Flask==3.0.0
flask-sock==0.7.0
flask-cors==4.0.0
Werkzeug==3.0.1

# ML Dependencies (RandomForest + MediaPipe)
scikit-learn==1.3.2
mediapipe==0.10.8
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# Text-to-Speech
pyttsx3==2.90

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Environment variables
python-dotenv==1.0.0